requirements.txt
blinker==1.9.0
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
Flask==3.1.2
fonttools==4.61.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
numpy==2.4.2
packaging==26.0
pandas==3.0.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
pycparser==3.0
pyparsing==3.3.2
pypdfium2==5.4.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.3
Werkzeug==3.1.5
